Finished the Writer and tons of bug fixes

+++ src/domain/model/Movie.java
@@ -43,7 +43,7 @@ public class Movie {
 
 	private Announcement movieAnnouncement;
 
-	public Movie(String title, String genre, int year, String director, int movieLength, double rating, String poster, double p, String syn) {
+	public Movie(String title, String genre, int year, String director, double movieLength, double rating, String poster, double p, String syn) {
 		this.movieID = movieCounter++;
 		this.title = title;
 		this.genre = genre;
@@ -57,7 +57,7 @@ public class Movie {
 		this.movieAnnouncement = null;
 	}
 
-	public Movie(int id,String title, String genre, int year, String director, int movieLength, double rating, String poster, double p, String syn) {
+	public Movie(int id,String title, String genre, int year, String director, double movieLength, double rating, String poster, double p, String syn) {
 		this.movieID = id;
 		movieCounter++;
 		this.title = title;
